Refactor code to remove circular reference in GameConfig
GameConfig has a list of GameRecords, but currently GameRecord is passed a GameConfig in order to calculate its own achievement. However, this creates a circular reference error when using Gson. Fix by replacing GameConfig parameter with int poorScore and int greatScore, which is the only data it needs to know anyway.